+++ src/exercises/chapter_5_linked_lists/circular_linked_list/Link.java
@@ -1,10 +1,10 @@
 package exercises.chapter_5_linked_lists.circular_linked_list;
 
 public class Link {
-    public int data;
+    public long data;
     public Link next;
 
-    public Link(int data) {
+    public Link(long data) {
         this.data = data;
     }
